pyproject.toml
[project]
name = "nimepa-dp-generator"
version = "1.0.0"
description = "DP Generator for NIMEPA Beach Cleanup Flyer"
requires-python = ">=3.8"
dependencies = [
    "rembg[gpu]>=2.0.50",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
